Cast to unsigned to avoid sign-extension problems.

+++ sys/libkern/ffsl.c
@@ -47,6 +47,6 @@ ffsl(long mask)
 	if (mask == 0)
 		return (0);
 	for (bit = 1; !(mask & 1); bit++)
-		mask >>= 1;
+		(unsigned long)mask >>= 1;
 	return (bit);
 }
